#122 Remove IEntity Dependency.
Creates new Root classes so we can inherit the basics without having to have an IEntity object.
