+++ uSync8.BackOffice/SyncHandlers/SyncHandlerContainerBase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+
 using Umbraco.Core.Cache;
 using Umbraco.Core.Events;
 using Umbraco.Core.Logging;
@@ -39,24 +40,13 @@ namespace uSync8.BackOffice.SyncHandlers
         protected SyncHandlerContainerBase(
             IEntityService entityService,
             IProfilingLogger logger,
-            ISyncSerializer<TObject> serializer,
-            ISyncTracker<TObject> tracker,
-            AppCaches appCaches,
-            SyncFileService syncFileService)
-            : base(entityService, logger, serializer, tracker, appCaches, syncFileService)
-        { }
-
-        protected SyncHandlerContainerBase(
-            IEntityService entityService,
-            IProfilingLogger logger,
             AppCaches appCaches,
             ISyncSerializer<TObject> serializer,
             SyncTrackerCollection trackers,
             SyncDependencyCollection checkers,
             SyncFileService syncFileService)
             : base(entityService, logger, appCaches, serializer, trackers, checkers, syncFileService)
-        {
-        }
+        { }
 
         [Obsolete("Construct your handler using the tracker & Dependecy collections for better checker support")]
         protected SyncHandlerContainerBase(
@@ -68,9 +58,18 @@ namespace uSync8.BackOffice.SyncHandlers
             ISyncDependencyChecker<TObject> checker,
             SyncFileService fileService)
             : base(entityService, logger, serializer, tracker, appCaches, checker, fileService)
-        {
+        { }
 
-        }
+        [Obsolete("Handler should take tracker and dependency checkers for completeness.")]
+        protected SyncHandlerContainerBase(
+            IEntityService entityService,
+            IProfilingLogger logger,
+            ISyncSerializer<TObject> serializer,
+            ISyncTracker<TObject> tracker,
+            AppCaches appCaches,
+            SyncFileService syncFileService)
+            : base(entityService, logger, serializer, tracker, appCaches, syncFileService)
+        { }
 
 
         protected IEnumerable<uSyncAction> CleanFolders(string folder, int parent)
@@ -81,11 +80,12 @@ namespace uSync8.BackOffice.SyncHandlers
             {
                 actions.AddRange(CleanFolders(folder, fdlr.Id));
 
-                if (!HasChildren(fdlr.Id))
+                if (!HasChildren(fdlr))
                 {
                     // get the name (from the slim)
                     var name = fdlr.Id.ToString();
-                    if (fdlr is IEntitySlim slim) {
+                    if (fdlr is IEntitySlim slim)
+                    {
                         name = slim.Name;
                     }
