+++ uSync8.Core/Tracking/SyncBaseTracker.cs
@@ -20,7 +20,6 @@ using uSync8.Core.Serialization;
 namespace uSync8.Core.Tracking
 {
     public abstract class SyncBaseTracker<TObject>
-        where TObject : IEntity
     {
         protected readonly ISyncSerializer<TObject> serializer;
