+++ uSync8.Core/uSync8.Core.csproj
@@ -190,6 +190,7 @@
     <Compile Include="Serialization\SyncSerializerAttribute.cs" />
     <Compile Include="Serialization\SyncSerializerBase.cs" />
     <Compile Include="Serialization\SyncSerializerOptions.cs" />
+    <Compile Include="Serialization\SyncSerializerRoot.cs" />
     <Compile Include="Serialization\SyncTreeSerializerBase.cs" />
     <Compile Include="Models\SyncAttempt.cs" />
     <Compile Include="Serialization\SyncSerializerCollectionBuilder.cs" />
